create home page on main.py

+++ src/main.py
@@ -1,4 +1,5 @@
 from textnode import TextNode
+from template import generate_page
 
 from os import path, listdir, mkdir
 
@@ -6,7 +7,8 @@ from shutil import copy, rmtree
 
 def reset_public_folder():
     public_folder_path = "public"
-    rmtree(public_folder_path)
+    if path.exists(public_folder_path):
+        rmtree(public_folder_path)
     print(f"Deleted '{public_folder_path}'")
     mkdir(public_folder_path)
     print(f"Created folder '{public_folder_path}'")
@@ -23,9 +25,13 @@ def copy_folder_contents(target_folder = "static", destination_folder = "public"
         else:
             copy(f"{target_folder}/{item}", f"{destination_folder}/{item}")
             print(f"Copied {target_folder}/{item} to {destination_folder}/{item}")
-    
+
+def generate_home_page():
+    generate_page('content/index.md', 'template.html', 'static/index.html')
+
 def main():
     reset_public_folder()
+    generate_home_page()
     copy_folder_contents()
 
 main()
